requirements.txt
Flask==2.0.2
Flask_Cors==3.0.9
Pillow
onnx
onnxruntime
